requirements.txt
# Core dependencies
apache-airflow==2.8.1
psycopg2-binary>=2.9.11,<3.0.0
httpx>=0.28.1,<0.29.0
networkx>=3.6.1,<4.0.0
redis>=7.2.0,<8.0.0
great-expectations>=1.12.3,<2.0.0
alembic>=1.18.4,<2.0.0
python-dotenv>=1.2.1,<2.0.0
dvc>=3.66.1,<4.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
google-cloud-storage>=2.14.0

# Dev dependencies
pytest>=9.0.2,<10.0.0
pytest-cov>=7.0.0,<8.0.0
pytest-asyncio>=1.3.0,<2.0.0
pytest-mock>=3.15.1,<4.0.0
ruff>=0.15.1,<0.16.0
mypy>=1.19.1,<2.0.0
pre-commit>=4.5.1,<5.0.0
black==24.10.0
